+++ test/integration/misc/user.js
@@ -29,22 +29,41 @@ const doUtil = require('../../../src/lib/doUtil');
 
 describe('User', function User() {
     this.timeout(600000);
+    let bigIPInstance;
 
     const key1 = 'ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQCwHJLJY+/U/ioAAAADAQABAAACAQCwHJLJY+z0Rb85in7Ean6JS2J9dzo1nSssm7ZyQvGgc1e7EVtztbVpHThsvw92+1hx9wlSogXN6Co5zrtqlN8/mvlQkRRQ+sp2To8PcSMeEVI+TqBOg6BWbwwNQLz9/eUJq2o4vBfSpsn7GSDIf6C3F9EahRPGCR/z0kw5GZob3Q== key1';
     const key2 = 'ssh-rsa AAAAB3NzaC1yc2EAu2Gr14xRiVLnG8KxNp2fO1/U/ioAz0Rb85in7Ean6JS2J9dzo1nSssm7ZyQvGgc1e7EVtztbVpHThsvw92+/mvlQkRRQ+sp2To8PcSMeEVI+TqBOg6BWbwwNQLzu2Gr14xRiVLnG8KxNp2fO19/eUJq2o4vBfSpsn7GSDIf6C3F9EahRPGCR/z0kw5GZob3Q== key2';
 
-    function getRootKeys() {
-        const catCmd = 'cat /root/.ssh/authorized_keys';
-        return doUtil.getBigIp(
-            undefined,
-            {
-                host: process.env.DO_HOST,
-                user: process.env.DO_USERNAME,
-                password: process.env.DO_PASSWORD
-            }
-        )
-            .then((bigIp) => doUtil.executeBashCommandIControl(bigIp, catCmd))
+    function getCommandResult(cmd) {
+        return Promise.resolve()
+            .then(() => {
+                if (!bigIPInstance) {
+                    bigIPInstance = doUtil.getBigIp(
+                        undefined,
+                        {
+                            host: process.env.DO_HOST,
+                            user: process.env.DO_USERNAME,
+                            password: process.env.DO_PASSWORD
+                        }
+                    );
+                }
+                return bigIPInstance;
+            })
+            .then((bigIp) => doUtil.executeBashCommandIControl(bigIp, cmd))
             .then((keyFile) => keyFile.trim());
+    }
+
+    function getRootKeys() {
+        return getCommandResult('cat /root/.ssh/authorized_keys');
+    }
+
+    function getDisableRootLogin() {
+        return getCommandResult('tmsh list sys db systemauth.disablerootlogin')
+            .then((response) => {
+                const keyVals = response.split(/\s+/);
+                const disableRootLoginValue = keyVals[keyVals.indexOf('{') + 2];
+                return disableRootLoginValue.trim();
+            });
     }
 
     it('userType root', () => {
@@ -97,5 +116,71 @@ describe('User', function User() {
                 assert.strictEqual(keys.length, 1);
                 assert.ok(keys[0].endsWith, 'Host Processor Superuser');
             });
+    });
+
+    it('disableRootLogin for root user', () => {
+        const decl = {
+            async: true,
+            class: 'Device',
+            schemaVersion: '1.24.0',
+            Common: {
+                class: 'Tenant',
+                root: {
+                    class: 'User',
+                    userType: 'root',
+                    disableRootLogin: true
+                }
+            }
+        };
+
+        const decl1 = {
+            async: true,
+            class: 'Device',
+            schemaVersion: '1.24.0',
+            Common: {
+                class: 'Tenant',
+                root: {
+                    class: 'User',
+                    userType: 'root',
+                    oldPassword: process.env.DO_PASSWORD,
+                    newPassword: process.env.DO_PASSWORD,
+                    disableRootLogin: false
+                }
+            }
+        };
+
+        const logInfo = {
+            declarationIndex: 0
+        };
+
+        return Promise.resolve()
+            .then(() => getDisableRootLogin())
+            .then((disableRootLogin) => {
+                assert.strictEqual(disableRootLogin, '"false"');
+            })
+            .then(() => assert.isFulfilled(
+                postDeclaration(decl, { logInfo })
+            ))
+            .then((response) => {
+                assert.strictEqual(response.result.code, 200);
+            })
+            .then(() => getDisableRootLogin())
+            .then((disableRootLogin) => {
+                assert.strictEqual(disableRootLogin, '"true"');
+            })
+            .catch((error) => {
+                console.log(`Error ${JSON.stringify(error)}`);
+            })
+            .then(() => assert.isFulfilled(
+                postDeclaration(decl1, { logInfo })
+            ))
+            .then((response) => {
+                assert.strictEqual(response.result.code, 200);
+            })
+            .then(() => getDisableRootLogin())
+            .then((disableRootLogin) => {
+                assert.strictEqual(disableRootLogin, '"false"');
+            })
+            .finally(() => postDeclaration(decl1, { logInfo }));
     });
 });